backend/pyproject.toml
[project]
name = "automated-show-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.11",
]
